moved benefit cards to view
